fixed wildcard env var name and added logic to skip live certs copy

+++ create-update-upload-certs.sh
@@ -4,10 +4,10 @@ CERT_REGISTRATION_EMAIL=${CERT_REGISTRATION_EMAIL:-digital@sba.gov}
 CERT_RENEW_ONLY=${CERT_RENEW_ONLY:-false}
 LETS_ENCRYPT_DIRECTORY=/etc/letsencrypt
 
-# Wildcard options
-CERT_REGISTER_WILDARD=${CERT_REGISTER_WILDARD:-true}
-CERT_REGISTER_WILDARD_CMD=""
-[ "${CERT_REGISTER_WILDARD}" == "true" ] && CERT_REGISTER_WILDARD_CMD="-d *.${CERT_HOSTNAME}"
+# Wildcard optionsC
+CERT_REGISTER_WILDCARD=${CERT_REGISTER_WILDCARD:-true}
+CERT_REGISTER_WILDCARD_CMD=""
+[ "${CERT_REGISTER_WILDCARD}" == "true" ] && CERT_REGISTER_WILDCARD_CMD="-d *.${CERT_HOSTNAME}"
 
 # Dry run options
 DRY_RUN=${DRY_RUN:-false}
@@ -48,15 +48,20 @@ if [ -d "$LETS_ENCRYPT_DIRECTORY/archive/${CERT_HOSTNAME}" ] && [ ! -L "$LETS_EN
   certbot renew ${DRY_RUN_CMD}
 else
   echo "Archive directory [$LETS_ENCRYPT_DIRECTORY/archive/${CERT_HOSTNAME}] not found, creating new cert"
-  certbot certonly ${DRY_RUN_CMD} -n --agree-tos --email "${CERT_REGISTRATION_EMAIL}" --dns-route53 -d "${CERT_HOSTNAME}" ${CERT_REGISTER_WILDARD_CMD}
+  certbot certonly ${DRY_RUN_CMD} -n --agree-tos --email "${CERT_REGISTRATION_EMAIL}" --dns-route53 -d "${CERT_HOSTNAME}" ${CERT_REGISTER_WILDCARD_CMD}
 fi
 
 # Copy live certs to latest directory
-mkdir -p $LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}
-cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/privkey.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/privatekey.pem"
-cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/cert.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/publiccert.pem"
-cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/fullchain.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/cachainfull.pem"
-cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/chain.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/cachainshort.pem"
+if [ "${DRY_RUN}" == "false" ];
+then
+  mkdir -p $LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}
+  cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/privkey.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/privatekey.pem"
+  cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/cert.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/publiccert.pem"
+  cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/fullchain.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/cachainfull.pem"
+  cp -a "$LETS_ENCRYPT_DIRECTORY/live/${CERT_HOSTNAME}/chain.pem" "$LETS_ENCRYPT_DIRECTORY/latest/${CERT_HOSTNAME}/cachainshort.pem"
 
-# Sync local to s3
-aws s3 sync --no-progress --exclude "live/*" $LETS_ENCRYPT_DIRECTORY/ "s3://${CERT_BUCKET_NAME}${CERT_BUCKET_PATH}"
+  # Sync local to s3
+  aws s3 sync --no-progress --exclude "live/*" $LETS_ENCRYPT_DIRECTORY/ "s3://${CERT_BUCKET_NAME}${CERT_BUCKET_PATH}"
+else
+  echo "Skipping cert copy and sync because this is a dry run"
+fi
